requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
wikipedia==1.4.0
python-dateutil==2.8.2
prometheus-client==0.19.0
mlflow==2.9.2
python-dotenv==1.0.0
pyyaml==6.0.1